requirements.txt
streamlit
praw
pandas
plotly
google.genai